refactor: decompose monolithic CameraRender into focused rendering components
Split the ~1800-line CameraRender god class into five single-responsibility
components, each owning one stage of the camera rendering pipeline:

- CameraPlaneRenderer  — plane geometry, texture assembly, mask application, and wall/floor/landscape classification
- CameraSpriteRenderer — sprite drawing (avatars, palettes, recoloring, alpha blending, bounds queries)
- CameraDoorwayRenderer — doorway masking, avatar cluster detection, occlusion ordering, and in-doorway avatar reveals
- CameraImageLoader    — sprite file resolution with path-traversal hardening and bounded URL fetching with SSRF guards
- CameraRenderUtils    — shared pixel utilities, alpha bounds, mask expansion, and rendering hints

CameraRender itself now orchestrates the pipeline: plane/sprite depth
sorting, filter/effect application, and background fill. The doorway
occlusion pass (previously inlined) is fully extracted and now correctly
handles multi-avatar clusters, foreground floor planes, and wall-top-cap
colour matching.

Tests are updated to match the new rendering expectations — most notably
avatars outside the doorway opening are now correctly occluded by the
wall (previously they leaked through), the wall-top-cap test now expects
the plane's own colour (not the wall colour), and a new test confirms
path-traversal rejection in CameraImageLoader.
